requirements.txt
django>=3.1.3,<3.2.0
Pillow>=8.0.1,<8.2.0
django-emoji-picker>=0.0.6,<0.1.0